+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [2.5.0] - 2023-08-23
 - Upgraded python-lib to v2.10.1
+- Replaced the arg_parser code with gen_class.ArgParser code.
 
 ### Changed
+- Multiple functions: Replaced the arg_parser code with gen_class.ArgParser code.
 - main: Removed gen_libs.get_inst call.